refactor and fix some stuff
