--- server/auth/reputation.go
@@ -1,175 +1,0 @@
-// This code is available on the terms of the project LICENSE.md file,
-// also available online at https://blueoakcouncil.org/license/1.0.0.
-
-package auth
-
-import (
-	"math"
-	"time"
-
-	"decred.org/dcrdex/dex/encode"
-	"decred.org/dcrdex/dex/order"
-)
-
-const (
-	cancelThreshWindow = 100 // spec
-	scoringMatchLimit  = 60  // last N matches (success or at-fault fail) to be considered in swap inaction scoring
-	scoringOrderLimit  = 40  // last N orders to be considered in preimage miss scoring
-
-	// preimage miss
-	preimageMissScore = -2 // book spoof, no match, no stuck funds
-
-	// failure to act violations
-	noSwapAsMakerScore   = -4  // book spoof, match with taker order affected, no stuck funds
-	noSwapAsTakerScore   = -11 // maker has contract stuck for 20 hrs
-	noRedeemAsMakerScore = -7  // taker has contract stuck for 8 hrs
-	noRedeemAsTakerScore = -1  // just dumb, counterparty not inconvenienced
-
-	successScore = 1 // offsets the violations
-
-	DefaultBaselineScore = 20
-
-	MaxScore = DefaultBaselineScore + scoringMatchLimit + scoringOrderLimit
-)
-
-// Violation represents a specific infraction. For example, not broadcasting a
-// swap contract transaction by the deadline as the maker.
-type Violation int32
-
-const (
-	ViolationInvalid Violation = iota - 2
-	ViolationForgiven
-	ViolationSwapSuccess
-	ViolationPreimageMiss
-	ViolationNoSwapAsMaker
-	ViolationNoSwapAsTaker
-	ViolationNoRedeemAsMaker
-	ViolationNoRedeemAsTaker
-)
-
-var violations = map[Violation]struct {
-	score int32
-	desc  string
-}{
-	ViolationSwapSuccess:     {successScore, "swap success"},
-	ViolationForgiven:        {-1, "forgiveness"},
-	ViolationPreimageMiss:    {preimageMissScore, "preimage miss"},
-	ViolationNoSwapAsMaker:   {noSwapAsMakerScore, "no swap as maker"},
-	ViolationNoSwapAsTaker:   {noSwapAsTakerScore, "no swap as taker"},
-	ViolationNoRedeemAsMaker: {noRedeemAsMakerScore, "no redeem as maker"},
-	ViolationNoRedeemAsTaker: {noRedeemAsTakerScore, "no redeem as taker"},
-	ViolationInvalid:         {0, "invalid violation"},
-}
-
-// Score returns the Violation's score, which is a representation of the
-// relative severity of the infraction.
-func (v Violation) Score() int32 {
-	return violations[v].score
-}
-
-// String returns a description of the Violation.
-func (v Violation) String() string {
-	return violations[v].desc
-}
-
-// NoActionStep is the action that the user failed to take. This is used to
-// define valid inputs to the Inaction method.
-type NoActionStep uint8
-
-const (
-	SwapSuccess NoActionStep = iota // success included for accounting purposes
-	NoSwapAsMaker
-	NoSwapAsTaker
-	NoRedeemAsMaker
-	NoRedeemAsTaker
-)
-
-// Violation returns the corresponding Violation for the misstep represented by
-// the NoActionStep.
-func (step NoActionStep) Violation() Violation {
-	switch step {
-	case SwapSuccess:
-		return ViolationSwapSuccess
-	case NoSwapAsMaker:
-		return ViolationNoSwapAsMaker
-	case NoSwapAsTaker:
-		return ViolationNoSwapAsTaker
-	case NoRedeemAsMaker:
-		return ViolationNoRedeemAsMaker
-	case NoRedeemAsTaker:
-		return ViolationNoRedeemAsTaker
-	default:
-		return ViolationInvalid
-	}
-}
-
-// String returns the description of the NoActionStep's corresponding Violation.
-func (step NoActionStep) String() string {
-	return step.Violation().String()
-}
-
-// Reputation is the user's reputation, which takes into account recent orders
-// and matches.
-type Reputation struct {
-	baseline int32
-	matches  *latestMatchOutcomes
-	ords     *latestPreimageOutcomes
-}
-
-// NewReputation is the constructor for a *Reputation.
-func NewReputation(baseline int32) *Reputation {
-	return &Reputation{
-		baseline: baseline,
-		matches:  newLatestMatchOutcomes(scoringMatchLimit),
-		ords:     newLatestPreimageOutcomes(scoringOrderLimit),
-	}
-}
-
-// RawScore is the score in the internal scoring system. RawScore will be on the
-// range [0, MaxScore]. The RawScore is computed from the user's recent match
-// outcomes and preimage history.
-func (r *Reputation) RawScore() (score int32) {
-	score = r.baseline
-	for v, count := range r.matches.binViolations() {
-		score += v.Score() * int32(count)
-	}
-	score += ViolationPreimageMiss.Score() * r.ords.misses()
-	return
-}
-
-// Score is the normalized user score, and will range from < 0 (for penalized
-// users) to max 100.
-func (r *Reputation) Score() int {
-	return int(math.Round(float64(r.RawScore()) / MaxScore * 100))
-}
-
-// Privilege is the amount of privilege the user has, and is computed as
-// (rawScore - baseline) / (maxScore - baseline), with a minimum value of zero.
-func (r *Reputation) Privilege() float64 {
-	rawScore := r.RawScore()
-	if rawScore < r.baseline {
-		return 0
-	}
-	return float64(rawScore-r.baseline) / float64(MaxScore-r.baseline)
-}
-
-// registerMatchOutcome registers the outcome from a match.
-func (r *Reputation) registerMatchOutcome(violation Violation, base, quote uint32, matchID order.MatchID, value uint64, refTime time.Time) {
-	r.matches.add(&matchOutcome{
-		time:    encode.UnixMilli(refTime),
-		mid:     matchID,
-		outcome: violation,
-		value:   value,
-		base:    base,
-		quote:   quote,
-	})
-}
-
-// registerMatchOutcome registers the outcome from a preimage request.
-func (r *Reputation) registerPreimageOutcome(miss bool, oid order.OrderID, refTime time.Time) {
-	r.ords.add(&preimageOutcome{
-		time: encode.UnixMilli(refTime),
-		oid:  oid,
-		miss: miss,
-	})
-}
